Attempt to write meta tags for scraping by FB etc
Wrote the tags for FB, Google, Twitter when I showImg. Confirmed through
Chrome Developer Tools that meta tags are being written.

BUT Facebook, and perhaps other sites, don't run Javascript in the pages
they scrape.

Need to PHP script to check if user-agent is Facebook etc, and if so
respond with a simple HTML-only page with the required meta tags.

Saving this version so I can reuse the code in the PHP script.